requirements.txt
python-telegram-bot[job-queue]==21.3
google-generativeai==0.8.3
aiosqlite==0.20.0
apscheduler==3.10.4
feedparser==6.0.11
httpx==0.27.0
python-dotenv==1.0.1
